+++ Src/DfT.DTRO.ApiTests/ApiTests/Scenarios/Schema_3.4.0/3.4.0_ConsumerScenarios/3.4.0_HappyScenarios.cs
@@ -21,7 +21,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.ConsumerScenarios
             // Prepare DTRO
             string dtroCreationJson = fileName
                                     .GetJsonFromFile(schemaVersionToTest)
-                                    .ModifyTraInDtroJson(schemaVersionToTest, publisher.TraId);
+                                    .ModifyTraInDtroJson(schemaVersionToTest, (int)publisher.TraId);
 
             // Send DTRO
             HttpResponseMessage dtroCreationResponse = await dtroCreationJson.SendJsonInDtroCreationRequestAsync(publisher);
@@ -32,7 +32,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.ConsumerScenarios
             // Search for DTRO as consumer
             TestUser consumer = await TestUsers.GetUser(TestUserType.Consumer);
 
-            string searchRequestJson = Dtros.GetSearchRequestJson("traCreator", publisher.TraId);
+            string searchRequestJson = Dtros.GetSearchRequestJson("traCreator", (int)publisher.TraId);
             HttpResponseMessage dtroSearchResponse = await searchRequestJson.GetDtroSearchResponseAsync(consumer);
             string dtroSearchResponseJson = await dtroSearchResponse.Content.ReadAsStringAsync();
             Assert.True(HttpStatusCode.OK == dtroSearchResponse.StatusCode,
@@ -50,7 +50,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.ConsumerScenarios
             // Prepare DTRO
             string dtroCreationJson = fileName
                                     .GetJsonFromFile(schemaVersionToTest)
-                                    .ModifyTraInDtroJson(schemaVersionToTest, publisher.TraId);
+                                    .ModifyTraInDtroJson(schemaVersionToTest, (int)publisher.TraId);
 
             string dtroTempFilePath = dtroCreationJson.CreateDtroTempFile(fileName, publisher);
 
@@ -63,7 +63,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.ConsumerScenarios
             // Search for DTRO as consumer
             TestUser consumer = await TestUsers.GetUser(TestUserType.Consumer);
 
-            string searchRequestJson = Dtros.GetSearchRequestJson("traCreator", publisher.TraId);
+            string searchRequestJson = Dtros.GetSearchRequestJson("traCreator", (int)publisher.TraId);
             HttpResponseMessage dtroSearchResponse = await searchRequestJson.GetDtroSearchResponseAsync(consumer);
             string dtroSearchResponseJson = await dtroSearchResponse.Content.ReadAsStringAsync();
             Assert.True(HttpStatusCode.OK == dtroSearchResponse.StatusCode,
