+++ Src/DfT.DTRO.ApiTests/ApiTests/Scenarios/Schema_3.4.0/3.4.0_PublisherScenarios/3.4.0_DtroUpdateScenarios/3.4.0_HappyScenarios.cs
@@ -41,7 +41,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.PublisherScenarios.DtroUpdateS
             // Prepare DTRO
             string dtroCreationJson = fileName
                                     .GetJsonFromFile(schemaVersionToTest)
-                                    .ModifyTraInDtroJson(schemaVersionToTest, publisher.TraId);
+                                    .ModifyTraInDtroJson(schemaVersionToTest, (int)publisher.TraId);
 
             // Send DTRO
             HttpResponseMessage dtroCreationResponse = await dtroCreationJson.SendJsonInDtroCreationRequestAsync(publisher);
@@ -85,7 +85,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.PublisherScenarios.DtroUpdateS
             // Prepare DTRO
             string dtroCreationJson = fileName
                                     .GetJsonFromFile(schemaVersionToTest)
-                                    .ModifyTraInDtroJson(schemaVersionToTest, publisher.TraId);
+                                    .ModifyTraInDtroJson(schemaVersionToTest, (int)publisher.TraId);
 
             string tempFilePathForDtroCreation = dtroCreationJson.CreateDtroTempFile(fileName, publisher);
 
@@ -134,7 +134,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.PublisherScenarios.DtroUpdateS
             // Prepare DTRO
             string dtroCreationJson = createFileWithSchema3_3_1
                                     .GetJsonFromFile(oldSchemaVersion)
-                                    .ModifyTraInDtroJson(oldSchemaVersion, publisher.TraId);
+                                    .ModifyTraInDtroJson(oldSchemaVersion, (int)publisher.TraId);
 
             // Send DTRO
             HttpResponseMessage dtroCreationResponse = await dtroCreationJson.SendJsonInDtroCreationRequestAsync(publisher);
@@ -145,7 +145,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.PublisherScenarios.DtroUpdateS
             // Prepare DTRO update
             string dtroUpdateJson = updateFileWithSchema3_4_0
                                         .GetJsonFromFile(schemaVersionToTest)
-                                        .ModifyTraInDtroJson(schemaVersionToTest, publisher.TraId)
+                                        .ModifyTraInDtroJson(schemaVersionToTest, (int)publisher.TraId)
                                         .ModifySourceActionType(schemaVersionToTest, "amendment")
                                         .ModifyTroNameForUpdate(schemaVersionToTest);
 
@@ -181,7 +181,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.PublisherScenarios.DtroUpdateS
             // Prepare DTRO
             string dtroCreationJson = createFileWithSchema3_3_1
                                     .GetJsonFromFile(oldSchemaVersion)
-                                    .ModifyTraInDtroJson(oldSchemaVersion, publisher.TraId);
+                                    .ModifyTraInDtroJson(oldSchemaVersion, (int)publisher.TraId);
 
             string tempFilePathForDtroCreation = dtroCreationJson.CreateDtroTempFile(createFileWithSchema3_3_1, publisher);
 
@@ -194,7 +194,7 @@ namespace DfT.DTRO.ApiTests.ApiTests.Schema_3_4_0.PublisherScenarios.DtroUpdateS
             // Prepare DTRO update
             string dtroUpdateJson = updateFileWithSchema3_4_0
                                         .GetJsonFromFile(schemaVersionToTest)
-                                        .ModifyTraInDtroJson(schemaVersionToTest, publisher.TraId)
+                                        .ModifyTraInDtroJson(schemaVersionToTest, (int)publisher.TraId)
                                         .ModifySourceActionType(schemaVersionToTest, "amendment")
                                         .ModifyTroNameForUpdate(schemaVersionToTest);
